+++ frame/azure-client/src/blob/blob.rs
@@ -7,6 +7,7 @@ use azure_storage::core::clients::{StorageAccountClient, StorageClient};
 use bytes::Bytes;
 use reqwest;
 use std::sync::Arc;
+#[cfg(test)]
 use url::Url;
 
 /// BlobClient is to access the Azure Storage APIs
@@ -28,6 +29,7 @@ impl BlobClient {
     }
 
     /// new_emulator instantiates a BlobClient object that is for use Azurite only.
+    #[cfg(test)]
     pub fn new_emulator(
         blob_storage_url_str: impl Into<String>,
         table_storage_url_str: impl Into<String>,
@@ -119,5 +121,33 @@ impl BlobClient {
             .map_err(|err| anyhow!(err))?;
 
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::BlobClient;
+
+    #[tokio::test]
+    async fn test_blob() {
+        let client = BlobClient::new_emulator("http://127.0.0.1:10000", "http://127.0.0.1:10002");
+
+        // コンテナがなければ作成する失敗しても無視
+        let _res = client.create_container("devstoreaccount1/emulcont").await;
+
+        // emulcontコンテナが存在することを確認する
+        let res = client.list_containers().await.unwrap();
+        assert_eq!(vec!("emulcont"), res);
+
+        // blobにデータをputする
+        let data = "bbbbb";
+        let _res = client
+            .put("emulcont", "test.txt", data.as_bytes())
+            .await
+            .unwrap();
+
+        // putしたデータを取得できることを確認する
+        let res = client.get("emulcont", "test.txt").await.unwrap();
+        assert_eq!(data, res);
     }
 }
